% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioactivity.R
\name{assay_data}
\alias{assay_data}
\title{Retrieve assay data from environmental variables}
\usage{
assay_data()
}
\value{
data.frame of assay data.
}
\description{
Retrieve assay data from environmental variables
}
\examples{
\dontshow{if (FALSE) withAutoprint(\{ # examplesIf}
# Examine assay data
head(b)
\dontshow{\}) # examplesIf}
}
