% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/httk.R
\name{calc_internal_dose_td}
\alias{calc_internal_dose_td}
\title{Calculate time-dependent internal chemical dose}
\usage{
calc_internal_dose_td(C_ext, IR, BW, scaling = 1, timestep = NULL)
}
\arguments{
\item{C_ext}{Matrix of external exposure, with columns given by \code{time} and
\code{dose}. This is in units of \eqn{\frac{mg}{m^3}}}

\item{IR}{Inhalation rate in \eqn{\frac{m^3}{day}}.}

\item{BW}{Body weight in \eqn{kg}.}

\item{scaling}{scaling factor encompassing any required unit adjustments.}

\item{timestep}{Duration of each dose in days. If \code{NULL}, duration of doses
is determined by using differences in dose times from \code{C_ext} matrix, with
last dose determined by the mean value of dose duration intervals.}
}
\value{
Matrix of internal chemical doses in \eqn{\frac{mg}{kg}}.
}
\description{
Calculate time-dependent internal chemical dose
}
\examples{
\dontshow{if (FALSE) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
\dontshow{\}) # examplesIf}
}
