% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/httk.R
\name{combine_dosing_pathways}
\alias{combine_dosing_pathways}
\title{Combine dosing data from multiple pathways}
\usage{
combine_dosing_pathways(dosing_matrices = NULL)
}
\arguments{
\item{dosing_matrices}{List of matrices, each representing a time-series of
chemical exposure.}
}
\value{
A matrix with two columns, \code{time} and \code{dose}, which totals all the
chemical exposures across all time points in the time-series exposure data.
}
\description{
Combine dosing data from multiple pathways
}
\examples{
\dontshow{if (FALSE) withAutoprint(\{ # examplesIf}
a <- matrix(c(seq(from = 0, to = 1, by = .1), 20*exp(-1*0:10)),
 ncol = 2, dimnames = list(NULL, c('time', 'dose')))
b <- matrix(c(seq(from = 0, to = 1, by = 1/6), 1:7),
ncol = 2, dimnames = list(NULL, c('time', 'dose')))
c <- combine_dosing_pathways(dosing_matrices = list(a,b))
c
\dontshow{\}) # examplesIf}
}
