% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/httk.R
\name{compute_c_ext_sensitivity}
\alias{compute_c_ext_sensitivity}
\title{Sensitivity analysis for time-dependent external concentration}
\usage{
compute_c_ext_sensitivity(
  chem.cas = NULL,
  C_ext = NULL,
  model_parameters = NULL,
  model = "pbtk",
  hill_params = NULL,
  IR = NULL,
  max_mult = 1
)
}
\arguments{
\item{chem.cas}{A vector containing the CASRN for a single chemical or
several chemicals in a mixture.}

\item{C_ext}{A matrix with external chemical concentration time-series data.
The column names must either be \code{time} and \code{dose} in the case of a single
chemical or \code{time} and the CASRN for each chemical in a mixture of several
chemicals.}

\item{model_parameters}{A data.frame of model parameters supplied by the
functions \code{parametrize_httk()} representing individuals in a population.}

\item{model}{A character string indicating which \code{httk} model to use, from
the options 'pbtk', '1comp', '3comp', 'gas_pbtk', 'fetal_pbtk'. The default
is 'pbtk'.}

\item{hill_params}{A data.frame output from the function \code{fit_hill()}, with
data corresponding to the chemical(s) contained in the \code{chem.cas}
parameter.}

\item{IR}{The inhalation rate for the average person of the population given
by the \code{model_parameters} input, in \eqn{\frac{m^3}{day}}.}

\item{max_mult}{The upper bound multiplier for max response. See
\code{calc_concentration_response()} for more details.}
}
\value{
A data.frame consisting of mixture response data from one or several
chemicals. See \code{calc_concentration_response()} for more details. For a
mixture of chemicals, this will be the set of common chemicals contained in
the \code{chem.cas}, \code{C_ext}, and \code{hill_params} inputs.
}
\description{
Sensitivity analysis for time-dependent external concentration
}
\examples{
\dontshow{if (FALSE) withAutoprint(\{ # examplesIf}
\dontshow{\}) # examplesIf}
}
