% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/httk.R
\name{compute_time_resolution_sensitivity}
\alias{compute_time_resolution_sensitivity}
\title{Time resolution sensitivity analysis for time-dependent external
concentration}
\usage{
compute_time_resolution_sensitivity(
  chem.cas = NULL,
  C_ext = NULL,
  model_parameters = NULL,
  model = "pbtk",
  hill_params = NULL,
  IR = NULL,
  max_mult = 1,
  resolutions = NULL,
  sample = "left"
)
}
\arguments{
\item{chem.cas}{A vector containing the CASRN for a single chemical or
several chemicals in a mixture.}

\item{C_ext}{A matrix with external chemical concentration time-series data.
The column names must either be \code{time} and \code{dose} in the case of a single
chemical or \code{time} and the CASRN for each chemical in a mixture of several
chemicals.}

\item{model_parameters}{A data.frame of model parameters supplied by the
functions \code{parametrize_httk()} representing individuals in a population.}

\item{model}{A character string indicating which \code{httk} model to use, from
the options 'pbtk', '1comp', '3comp', 'gas_pbtk', 'fetal_pbtk'. The default
is 'pbtk'.}

\item{hill_params}{A data.frame output from the function \code{fit_hill()}, with
data corresponding to the chemical(s) contained in the \code{chem.cas}
parameter.}

\item{IR}{The inhalation rate for the average person of the population given
by the \code{model_parameters} input, in \eqn{\frac{m^3}{day}}.}

\item{max_mult}{The upper bound multiplier for max response. See
\code{calc_concentration_response()} for more details.}

\item{resolutions}{Time resolutions to subdivide the time-series data. These
are positive numeric values, with units in days.}

\item{sample}{The values in \code{resolutions} divide the entire time-series data
and \code{sample} determine which time point to use for each interval. The valid
options are 'left', 'right', 'center', with a default value of 'left'.}
}
\value{
A named list with entries corresponding to a data.frame consisting
of mixture response data from one or several chemicals. See
\code{compute_c_ext_sensitivity()} for more details on the structure of each
entry. The names are the valid time resolutions given by the \code{resolutions}
parameter.
}
\description{
Time resolution sensitivity analysis for time-dependent external
concentration
}
\examples{
\dontshow{if (FALSE) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
\dontshow{\}) # examplesIf}
}
