% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/httk.R
\name{generate_steady_state_css}
\alias{generate_steady_state_css}
\title{Generate steady state concentration (CSS)}
\usage{
generate_steady_state_css(
  data_input = NULL,
  n_samples = 500,
  load_all_httk_chemical_data = FALSE,
  set_seed = FALSE,
  seed = 42
)
}
\arguments{
\item{data_input}{A list with non-empty \code{dose_response} entry. This entry
should contain a data.frame with a column \code{casn}.}

\item{n_samples}{Number of sampled CSS values to create.}

\item{load_all_httk_chemical_data}{Boolean determining whether to load
additional \code{httk} chemical data.}

\item{set_seed}{Boolean for setting a seed.}

\item{seed}{Seed, used only if \code{set_seed} has value \code{TRUE}.}
}
\value{
A data.frame with generated CSS values for a combination of age and
weight parameters, including median values grouped by weight and by age.
}
\description{
Generate steady state concentration (CSS)
}
\examples{
\dontshow{if (FALSE) withAutoprint(\{ # examplesIf}
\dontshow{\}) # examplesIf}
}
