% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioactivity.R
\name{get_active_assays}
\alias{get_active_assays}
\title{Retrieve assay data}
\usage{
get_active_assays(assays = NULL, chemids = NULL, hitc = 0.9)
}
\arguments{
\item{assays}{A vector of assay names.}

\item{chemids}{A vector of DTXSID chemical identifiers}

\item{hitc}{A value between 0 and 1 indicating the cutoff level for assay
inclusion}
}
\value{
A processed tibble with columns \code{endp} for endpoint, \code{casn} for
CASRN, \code{dtxsid} for chemical DTXSID, \code{chnm} for chemical name,
\code{logc} for dose value, and \code{resp} for assay response.
}
\description{
Retrieve assay data
}
\examples{
\dontshow{if (FALSE) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
# Retrieve information for `TOX21_DT40` assay for chemicals with DTXSIDs
# `DTXSID5020865` and `DTXSID2045078`.
g <- get_active_assays(assays =  c('TOX21_DT40'),
chemids = c('DTXSID5020865', 'DTXSID2045078'))
g
\dontshow{\}) # examplesIf}
}
