% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/httk.R
\name{get_population_average}
\alias{get_population_average}
\title{Get population average}
\usage{
get_population_average(params = NULL)
}
\arguments{
\item{params}{Table of httk model parameters, output from}
}
\value{
Single row data.table of averages of columns
}
\description{
Get population average
}
\examples{
\dontshow{if (FALSE) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
\dontshow{\}) # examplesIf}
}
