% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/httk.R
\name{get_population_statistics}
\alias{get_population_statistics}
\title{Get population statistics}
\usage{
get_population_statistics(params = NULL, quantiles = NULL)
}
\arguments{
\item{params}{Table of httk model parameters, output from}

\item{quantiles}{A vector of quantile values, from 0 to 1. If not supplied,
the average of each column is returned.}
}
\value{
Single row data.table of statistics of columns
}
\description{
Get population statistics
}
\examples{
\dontshow{if (FALSE) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
\dontshow{\}) # examplesIf}
}
