% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/httk.R
\name{load_httk_data}
\alias{load_httk_data}
\title{Load \code{httk} data}
\usage{
load_httk_data()
}
\description{
This function loads in extended \code{httk} data and toggles the option
\code{httk_data_loaded} to \code{TRUE}.
}
\examples{
\dontshow{if (FALSE) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
\dontshow{\}) # examplesIf}
}
