% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/httk.R
\name{parametrize_httk}
\alias{parametrize_httk}
\title{Prepare parameters for a httk models}
\usage{
parametrize_httk(chem.cas, model = "pbtk", mcs = NULL)
}
\arguments{
\item{chem.cas}{The chemical identifier CASN}

\item{model}{For which \code{httk} model to prepare parameters, from the list
c('pbpk', '1comp', '3comp', 'gas_pbtk', 'fetal_pbtk', 'schmitt') with
default \code{pbtk}.}

\item{mcs}{data.frame output from \code{httk::create_mc_samples}.}
}
\value{
Model parameters for each row of the input \code{mcs} population
parameters
}
\description{
Prepare parameters for a httk models
}
\examples{
\dontshow{if (FALSE) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
\dontshow{\}) # examplesIf}
}
