% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/httk.R
\name{reset_httk_data}
\alias{reset_httk_data}
\title{Reset \code{httk} data}
\usage{
reset_httk_data()
}
\description{
Helper function that resets loaded \code{httk} data and toggles the option
\code{httk_data_loaded} to \code{FALSE}.
}
\examples{
\dontshow{if (FALSE) withAutoprint(\{ # examplesIf}
\dontshow{\}) # examplesIf}
}
