% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/response_decay.R
\name{response_decay}
\alias{response_decay}
\title{Exponential decay of response in time-varying dose-response}
\usage{
response_decay(
  plasma_data = NULL,
  max = NULL,
  AC50 = NULL,
  n = NULL,
  t_half = NULL,
  a = 1,
  b = 1
)
}
\arguments{
\item{plasma_data}{A data.frame containing dose-response data. Must contain
columns \code{time} and \code{Cplasma}.}

\item{max}{The max response level for assay.}

\item{AC50}{AC50 level for assay.}

\item{n}{The slope factor for the assay response.}

\item{t_half}{Half-life in days of response decay.}

\item{a}{Control parameter for new plasma level.}

\item{b}{Control parameter for decaying plasma level.}
}
\value{
A data.frame with a single column, \code{response}, that gives the
response due to new plasma readings and decaying plasma levels.
}
\description{
This function models time-varying dose-response given a time series of blood
plasma concentration, parameters associated to the dose-response curve, a
response decay rate, and optional weighting parameters for balancing the
decaying plasma levels with new plasma levels.
}
\details{
For a given dose-response relationship, let \eqn{r = f(d)} denote the response
level corresponding to a dose, with \eqn{f} the dose-response function. In this
framework, we combine the response from the previous time step with the new
dose to determine the next response level.

Let \eqn{r_t} denote the response at time \eqn{t}, \eqn{d_{t+\Delta t}} the
new plasma level at time \eqn{t + \Delta t}, and \eqn{r_{t+\Delta t}} the
response at time \eqn{t + \Delta t} and \eqn{t_{\frac{1}{2}}} the half-life
for the response decay. The decay at time \eqn{t + \Delta t} from the
response at time \eqn{t} is given by
\eqn{r^{d}_{t + \Delta t} = r_t\times\exp(-\frac{\ln(2)\times\Delta
t}{t_{\frac{1}{2}}})} and the corresponding dose is given by
\eqn{f^{-1}(r^d_{t + \Delta t})}. This is combined with the new plasma
reading, \eqn{d_{t + \Delta t}} as
\eqn{a\times d_{t + \Delta t} + b\times f^{-1}(r^d_{t + \Delta t})} and input
into \eqn{f} to determine the new response,
\eqn{r_{t + \Delta t} = f(a\times d_{t + \Delta t} + b\times f^{-1}(r^d_{t + \Delta t}))}.
}
\examples{
\dontshow{if (FALSE) withAutoprint(\{ # examplesIf}
\dontshow{\}) # examplesIf}
}
