% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/response_decay.R
\name{response_decay_exponential}
\alias{response_decay_exponential}
\title{Pre-equilibrium dose-response}
\usage{
response_decay_exponential(
  plasma_data = NULL,
  max = NULL,
  AC50 = NULL,
  n = NULL,
  K_decay = NULL,
  k_off = NULL,
  k_on = NULL
)
}
\arguments{
\item{plasma_data}{A data.frame containing plasma concentration data. Must
contain columns \code{time} (in days) and \code{Cplasma} (in \eqn{\mu}M).}

\item{max}{The max response level for assay.}

\item{AC50}{AC50 level for assay.}

\item{n}{The slope factor for the assay response.}

\item{K_decay}{Response decay constant in units \eqn{s^{-1}}.}

\item{k_off}{Dissociation constant in units \eqn{s^{-1}}}

\item{k_on}{Association constant in units \eqn{M^{-1}s^{-1}}}
}
\value{
A data.frame with a single column, \code{response}, that gives the
response due to new plasma readings and decaying plasma levels.
}
\description{
This models pre-equilibrium dose-response to a time-series of blood plasma
chemical concentration data. Users supply the plasma concentration
time-series data and either a generic response decay constant or dissociation
and association constants. In the former case, the decay depends only on the
time while in the latter, it also depends on the dose. The equation governing
the dose-response decay is given by \eqn{h(L, t) = \text{max}\frac{L^n}{L^n +
\text{AC50}^n}(1 - e^{-Kt})}, with \code{max} the maximal response and where the
constant \eqn{K} is either \code{K_decay} in the dose-independent setting or
\eqn{K = \text{k}_{\text{off}} + \text{k}_{\text{on}}L} in the
dose-dependent setting.
}
\examples{
\dontshow{if (FALSE) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
\dontshow{\}) # examplesIf}
}
