% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/response_decay.R
\name{response_decay_ode}
\alias{response_decay_ode}
\title{ODE version of Response Decay}
\usage{
response_decay_ode(
  plasma_data = NULL,
  max = NULL,
  min = NULL,
  AC50 = NULL,
  n = NULL,
  t_ss = NULL,
  tol = 1e-04
)
}
\arguments{
\item{plasma_data}{A data.frame containing dose-response data. Must contain
columns \code{time} and \code{Cplasma}.}

\item{max}{The max response level for assay.}

\item{min}{The min response level for assay.}

\item{AC50}{AC50 level for assay.}

\item{n}{The slope factor for the assay response.}

\item{t_ss}{Time to steady-state of assay response within tolerance level
\code{tol}.}

\item{tol}{A value strictly between 0 and 1, giving the level determining how
close the response level is to steady-state value. More specifically, let
\eqn{R_{ss}} is the steady-state response corresponding to a specific
constant dose level and \eqn{t_{ss}} the time it takes to achieve
\eqn{|\frac{R(t)}{R_{ss}}| < 1 - \text{tol}} for \eqn{t > t_{ss}}.}
}
\value{
A \code{deSolve} object with columns 'time', 'response', and
'new_response'. The column 'new_response' is the incremental response value
from the dose at the time point while 'response' gives the overall
response.
}
\description{
Models time-dependent assay response given plasma concentration following the
ODE system \eqn{\frac{dR}{dt} = -KR + Kf(d(t))} where \eqn{R} is the
response, \eqn{d} is the dose function, \eqn{f} is the dose-response
function, and \eqn{K} is a decay rate coefficient. For constant dose regime
\eqn{d}, the steady-state response is given by \eqn{R = f(d)}. The value of
the decay rate constant \eqn{K} is determined by \eqn{K =
-\frac{\log(\text{tol})}{t_{ss}}} where \eqn{\text{tol}} is the tolerance
level and \eqn{t_{ss}} is the time it takes the response to converge to
\eqn{R_{ss}} under constant dosing regime within a specified tolerance level.
See the description of \code{tol} below for more details.
}
\examples{
\dontshow{if (FALSE) withAutoprint(\{ # examplesIf}
\dontshow{\}) # examplesIf}
}
