% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/httk.R
\name{simulate_css}
\alias{simulate_css}
\title{Generate population CSS}
\usage{
simulate_css(chem.cas, agelim_years, weight_category, samples, verbose = TRUE)
}
\arguments{
\item{chem.cas}{Chemical identifier CASN.}

\item{agelim_years}{Vector of length 2 denoting age range of population.}

\item{weight_category}{String with values \code{Normal} or \code{Obese}.}

\item{samples}{Number of CSS values to generate.}

\item{verbose}{Boolean determining whether to display progress report of
internal function operations.}
}
\value{
A list of CSS values based on input parameter values.
}
\description{
Generate population CSS
}
\examples{
\dontshow{if (FALSE) withAutoprint(\{ # examplesIf}
\dontshow{\}) # examplesIf}
}
