% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/httk.R
\name{simulate_population_variability_parameters}
\alias{simulate_population_variability_parameters}
\title{Generate population parameters for time-dependent \code{httk} models}
\usage{
simulate_population_variability_parameters(
  chem.cas,
  agelim_years = NULL,
  age_matrix = NULL,
  weight_category,
  samples,
  model = "pbtk",
  verbose = TRUE,
  set_seed = FALSE,
  seed = 42
)
}
\arguments{
\item{chem.cas}{Chemical identifier CASN.}

\item{agelim_years}{Vector of length 2 denoting age range of population.}

\item{age_matrix}{A matrix with columns \code{age} and \code{samples} to determine
number of individuals to sample from each age.}

\item{weight_category}{String with values \code{Normal} or \code{Obese}.}

\item{samples}{Number of CSS values to generate.}

\item{model}{This can be 'pbtk', '3compartment', '3compartmentss', or
'1compartment'. Default for this function is 'pbtk', See documentation for
\code{httk::create_mc_samples} for more detailed information.}

\item{verbose}{Boolean determining whether to display progress report of
internal function operations.}

\item{set_seed}{Boolean for setting a seed.}

\item{seed}{Seed, used only if \code{set_seed} has value \code{TRUE}.}
}
\value{
A table of population parameters to use in time-dependent \code{httk}
models. If \code{age_matrix} parameter is supplied, an additional column \code{age}
is included in the output.
}
\description{
Generate population parameters for time-dependent \code{httk} models
}
\examples{
\dontshow{if (FALSE) withAutoprint(\{ # examplesIf}
\dontshow{\}) # examplesIf}
}
