% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/response_decay.R
\name{threshold_exceedance}
\alias{threshold_exceedance}
\title{Determine the time and AUC of excess response}
\usage{
threshold_exceedance(thresholds, response_data)
}
\arguments{
\item{thresholds}{A named list or vector of threshold names and response
threshold values.}

\item{response_data}{A time-series of response data, with columns \code{time} and
\code{response}.}
}
\value{
A data.frame consisting of a column of valid threshold names, the
duration of time the response exceeded the threshold values, and the AUC
between the threshold value and response curve.
}
\description{
Given a response curve and response threshold(s), it is possible to determine
for how long the response exceeds a particular threshold and the AUC of the
excess response above that threshold. This function takes in a named vector
or list of threshold values and a time-series of response data output from
the function \code{\link[=response_decay]{response_decay()}}.
}
\examples{
\dontshow{if (FALSE) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
\dontshow{\}) # examplesIf}
}
