% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/httk.R
\name{solve_httk_model}
\alias{solve_httk_model}
\title{Solve time-dependant httk model}
\usage{
solve_httk_model(
  chem.cas,
  model = "pbtk",
  parameters = NULL,
  plot = FALSE,
  data.matrix = NULL,
  input.units = "mg/kg",
  output.units = "uM",
  timestep = NULL,
  IR = NULL,
  standardize = TRUE
)
}
\arguments{
\item{chem.cas}{The chemical identifier CASN.}

\item{model}{Which \code{httk} model to solve, from the list c('pbpk', '1comp',
'3comp', 'gas_pbtk', 'fetal_pbtk', 'schmitt') with default \code{pbtk}.}

\item{parameters}{Set of parameters to supply to the specified httk model.
Each row corresponds to a unique individual.}

\item{plot}{Boolean to determine whether to create a plot of blood plasma
concentration vs time.}

\item{data.matrix}{Alternate parameter to supply time and dose information,
with columns \code{time} and \code{dose}.}

\item{input.units}{Set of units for input dose. See documentation for \code{httk}
models for more information.}

\item{output.units}{Set of units for output dose. See documentation for
\code{httk} models for more information.}

\item{timestep}{Duration of each dose in days. If \code{NULL}, duration of doses
is determined by using differences in dose times from \code{C_ext} matrix, with
last dose determined by the mean value of dose duration intervals.}

\item{IR}{A single value or vector of inhalation rate values corresponding to
the individuals given by the \code{parameters} data.}

\item{standardize}{Standardize exposure by individual-specific \code{IR} and
bodyweight.}
}
\value{
A list of httk model output values for each iteration and an
optional \code{ggplot2} plot of the Cplasma vs time for each individual
represented by the \code{parameters} argument.
}
\description{
Solve time-dependant httk model
}
\examples{
\dontshow{if (FALSE) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
\dontshow{\}) # examplesIf}
}
